+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,10 +1,10 @@
 package com.techcourse.controller;
 
+import com.interface21.webmvc.servlet.mvc.controller.Controller;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
