+++ app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,10 +1,10 @@
 package com.techcourse.controller;
 
+import com.interface21.webmvc.servlet.mvc.controller.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 
 public class LoginViewController implements Controller {
