+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,10 +1,10 @@
 package com.techcourse.controller;
 
+import com.interface21.webmvc.servlet.mvc.controller.Controller;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 
 public class RegisterController implements Controller {
