+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,8 +1,8 @@
 package com.techcourse.controller;
 
+import com.interface21.webmvc.servlet.mvc.controller.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 
 public class RegisterViewController implements Controller {
